backend/requirements.txt
alembic==1.14.1
apscheduler==3.11.0
asyncpg==0.30.0
cryptography==44.0.2
fastapi==0.115.12
greenlet==3.3.2
httpx==0.28.1
pydantic-settings==2.8.1
psycopg2-binary==2.9.11
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
sqlalchemy==2.0.48
structlog==25.2.0
uvicorn[standard]==0.34.0
plyer==2.1.0
